+++ src/screens/Home/Components/CourseFilter/index.scss
