--- src/screens/Home/Components/Feed/Filter/index.css
@@ -1,55 +1,0 @@
-
-.sp-filter {
-  width: 100%;
-  margin-top: 1rem;
-  padding: .4em 20px 10px 30px;
-  position: sticky;
-  top: 50px;
-  z-index: 10;
-  background: linear-gradient(to bottom,white 90%,rgba(255, 255, 255, 0));
-}
-@media screen and (max-width: 800px) {
-  .sp-filter {
-    top: 0;
-    padding: 10px 0 !important;
-    position: relative;
-  }
-}
-
-.sp-filter .sp-search-bar {
-  border-bottom: 1px solid rgb(228, 228, 228);
-  background: white;
-}
-.ui.stackable.grid>.row>.column[id=sp-filter-search] {
-  padding-top: 0 !important;
-}
-.sp-filter .ui.selection.dropdown {
-  background: none !important;
-  border-top-left-radius: 5px !important;
-  border-top-right-radius: 5px !important;
-  border-color: transparent !important;
-}
-.sp-filter .ui.selection.active.dropdown .menu {
-  border-bottom-left-radius: 5px !important;
-  border-bottom-right-radius: 5px !important;
-}
- .ui.default.dropdown:not(.button)>.text, .ui.dropdown:not(.button)>.default.text {
-  color: rgb(100, 100, 100) !important;
-}
-.sp-filter .ui.dropdown .menu>.item {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  transition: all 0.2s linear;
-  -o-transition: all 0.2s linear;
-  -webkit-transition: all 0.2s linear;
-}
-.sp-filter .ui.dropdown .menu>.item:hover {
-  overflow: visible;
-  white-space: normal;
-}
-
-.sp-filter label {
-  color: rgb(37, 113, 116) !important;
-  padding-left: 1em;
-}
