--- src/screens/Home/Components/Feed/index.css
@@ -1,23 +1,0 @@
-/**
- * Stylesheet for Home screen
- */
-
-/* Content */
-
- .hide {
-  display: none;}
-  @keyframes fadeOut {
-    0% {opacity: 1;}
-    100% {opacity: 0;}}
-
-
-/* Offering list */
-
-.offering-list {
-  margin-top: 0rem;
-  padding: 0 15px;
-}
-
-
-  
-
