+++ src/screens/Home/Components/SectionList/index.scss
