+++ src/screens/Home/Components/index.js
@@ -1,1 +1,2 @@
-export { Feed } from './Feed';
+export { default as SectionList } from './SectionList';
+export { default as CourseFilter } from './CourseFilter';
