--- src/screens/Home/Utils/index.js
@@ -1,1 +1,0 @@
-export { generateHomeUserGuide } from './user-guide';
