+++ src/screens/Home/controllers/HomeConstants.js
