--- src/screens/Home/index.css
@@ -1,50 +1,0 @@
-/**
- * Stylesheet for student page/offering viewing page
- */
-
-.sp-bg {
-  -webkit-position: absolute;
-  -moz-position: absolute;
-  position: absolute;
-  z-index: 2;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow-y: auto;
-  overflow-x: hidden;
-  background-color: white;
-  -webkit-overflow-scrolling: touch;
-  font-family: Lato,Helvetica Neue,Arial,Helvetica,sans-serif;
-}
-
-.prompt {
-  border-radius: 3rem !important;
-}
-
-/* Side Bar */
-.op-sidebar .signin-prompt {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: 2rem;}
-  .op-sidebar .signin-prompt p {
-    font-size: 1.1rem;
-    text-align: center;
-    color: rgb(58, 122, 122);
-  }
-.op-sidebar .pico.pico-btn .pico-btn-con {
-  justify-content: center;
-}
-.goback-container {
-  width: 100%;
-  padding-bottom: 1rem;}
-.goback-container .del-icon {
-  transition: margin 0.2s linear;
-  -o-transition: margin 0.2s linear;
-  -webkit-transition: margin 0.2s linear;
-  color: rgb(48, 129, 129);
-  margin-right: 2rem;
-  cursor: pointer;
-  font-size: 1.2rem;}
-  .del-icon:hover, .del-icon:focus, .del-icon:active {
-    margin-left: 1rem;}
