--- src/screens/Home/transition.css
@@ -1,36 +1,0 @@
-/**
- * Stylesheet for animated transitions using react-transition-group
- */
-
-/* left in */
-.fade-enter {
-  opacity: 0;
-}
-.fade-enter-active {
-  opacity: 1;
-  transition: opacity 300ms, transform 300ms;
-  -o-transition: opacity 300ms, transform 300ms;
-  -webkit-transition: opacity 300ms, transform 300ms;
-}
-.fade-exit {
-  display: none;
-}
-
-
-
-/* top in */
-.sp-search-bar-enter {
-  opacity: 0;
-  transform: translateY(-40px) scaleX(1.03);
-  -o-transform: translateY(-40px) scaleX(1.03);
-  -webkit-transform: translateY(-40px) scaleX(1.03);
-}
-.sp-search-bar-enter-active {
-  opacity: 1;
-  transform: translateX(0) scaleX(1);
-  -o-transform: translateX(0) scaleX(1);
-  -webkit-transform: translateX(0) scaleX(1);
-  transition: opacity 300ms, transform 300ms;
-  -o-transition: opacity 300ms, transform 300ms;
-  -webkit-transition: opacity 300ms, transform 300ms;
-}
