--- src/screens/Home_v2/components/index.js
@@ -1,2 +1,0 @@
-export { default as SectionList } from './SectionList';
-export { default as CourseFilter } from './CourseFilter';
