+++ src/screens/index.js
@@ -12,8 +12,7 @@ export {
 } from './Instructor';
 export { MediaSettings } from './MediaSettings';
 
-// export { Home } from './Home';
-export { Home } from './Home_v2';
+export { Home } from './Home';
 export { Course } from './Course';
 export { Search } from './Search';
 export { History } from './History';
